1.3.9.0 [PUSH]
Pre-release version for staging Dalamud. Primarily for bug-testing or those who want the plugin ASAP. Expect issues from either the plugin or Dalamud itself. A 'stable' release will happen later, or this might be the stable release and the version number just gets bumped up, we will have to see.